imageBlock

+++ ShaderForRealityKit/S01_Glitch/GlitchRGBSplit.metal
@@ -161,7 +161,7 @@ void postProcessRGBSplitV4(uint2 gid [[thread_position_in_grid]],
     half2 inSize = half2(inColor.get_width(), inColor.get_height());
     float time = _TimeX * _Speed;
     // 计算抖动曲线
-    half splitAmount = _Indensity * randomNoise(time, 2);
+    half splitAmount = _Indensity * half(randomNoise(time, 2));
     
     // 计算分离后的坐标
     half2 offset = splitAmount * inSize;
